Move master-imported codes into a single test folder. Make sure to reinitialize truediv inferrer after running tests.infer.test_infer.test_specialization. Reformat code.
