+++ tests/infer/from_master/master_placeholders.py
