+++ tests/infer/from_master/test_operations.py
@@ -1,6 +1,7 @@
 from myia.testing.common import Bool, Nil, Ty
 from myia.testing.multitest import infer, mt
-from myia.testing.testing_inferrers import add_testing_inferrers
+
+from .master_inferrers import add_testing_inferrers
 
 add_testing_inferrers()
